[RTOP-76] Logging module
